requirements.txt
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
ansible==2.9.0rc5
ansible-runner==1.4.4
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
docutils==0.15.2
idna==2.8
Jinja2==2.10.3
lockfile==0.12.2
MarkupSafe==1.1.1
multidict==4.5.2
paramiko==2.6.0
pexpect==4.7.0
psutil==5.6.3
ptyprocess==0.6.0
pycparser==2.19
PyNaCl==1.3.0
python-daemon==2.2.4
PyYAML==5.1.2
six==1.12.0
yarl==1.3.0
